+++ CHANGELOG.md
@@ -1,3 +1,45 @@
+## [2.0.1](https://github.com/sanemat/do-not-merge-wip-for-github/compare/v2.0.0...v2.0.1) (2019-11-03)
+
+
+
+# [2.0.0](https://github.com/sanemat/do-not-merge-wip-for-github/compare/v1.0.10...v2.0.0) (2019-11-03)
+
+
+### Bug Fixes
+
+* **content:** detect issue label ([74921cc](https://github.com/sanemat/do-not-merge-wip-for-github/commit/74921cc3ec9e6a4f8baa911832cf41d93b5526d2))
+* **textcontent:** detect commit message ([f59c727](https://github.com/sanemat/do-not-merge-wip-for-github/commit/f59c7278dfe632a987b999308624754655109906)), closes [#204](https://github.com/sanemat/do-not-merge-wip-for-github/issues/204) [#236](https://github.com/sanemat/do-not-merge-wip-for-github/issues/236)
+
+
+### Features
+
+* **content:** drop branch protection ([ea8f3cd](https://github.com/sanemat/do-not-merge-wip-for-github/commit/ea8f3cd181e6a47819bbe1ba50135369c54852a1)), closes [#256](https://github.com/sanemat/do-not-merge-wip-for-github/issues/256)
+* **options:** drop branch protection ([25f289a](https://github.com/sanemat/do-not-merge-wip-for-github/commit/25f289a74855c7e09f76922264ee8e69f110f42a))
+
+
+
+<a name="1.0.10"></a>
+## [1.0.10](https://github.com/sanemat/do-not-merge-wip-for-github/compare/v1.0.9...v1.0.10) (2018-08-28)
+
+* Renew app icon ([974ec3d](https://github.com/sanemat/do-not-merge-wip-for-github/commit/974ec3df418631f68a1d2f443053b861367fbcc0))
+
+<a name="1.0.9"></a>
+## [1.0.9](https://github.com/sanemat/do-not-merge-wip-for-github/compare/v1.0.8...v1.0.9) (2017-10-10)
+
+
+### Features
+
+* disable the merge method button ([e066084](https://github.com/sanemat/do-not-merge-wip-for-github/commit/e066084))
+
+
+
+<a name="1.0.8"></a>
+## [1.0.8](https://github.com/sanemat/do-not-merge-wip-for-github/compare/v1.0.7...v1.0.8) (2017-08-24)
+
+* Add support for () and : ([5486c3e](https://github.com/sanemat/do-not-merge-wip-for-github/commit/5486c3ebd85f848d71f4593c1431d394cc797921))
+* Add support DONT and DON'T ([0ba90de](https://github.com/sanemat/do-not-merge-wip-for-github/commit/0ba90de6aa02c72f29a4cd95fb7442abb989d2d4))
+
+
 <a name="1.0.7"></a>
 ## [1.0.7](https://github.com/sanemat/do-not-merge-wip-for-github/compare/v1.0.6...v1.0.7) (2016-08-27)